fix(ansi): rename files to kebab-case for consistency
