+++ packages/cycle/src/presets/colors/ansi-all.ts
@@ -1,0 +1,4 @@
+import ansiBackgrounds from './ansi-backgrounds'
+import ansiForegrounds from './ansi-foregrounds'
+
+export default [...ansiForegrounds, ...ansiBackgrounds]
