+++ packages/cycle/src/presets/colors/ansi-foregrounds.ts
