--- packages/cycle/src/presets/colors/ansiAll.ts
@@ -1,4 +1,0 @@
-import ansiBackgrounds from './ansiBackgrounds'
-import ansiForegrounds from './ansiForegrounds'
-
-export default [...ansiForegrounds, ...ansiBackgrounds]
