+++ packages/cycle/src/presets/colors/index.ts
@@ -1,9 +1,9 @@
 // WARNING: This file is automatically generated. Any changes will be lost the next time the generator is run.
 
 export {default as ansi} from './ansi'
-export {default as ansiAll} from './ansiAll'
-export {default as ansiBackgrounds} from './ansiBackgrounds'
-export {default as ansiForegrounds} from './ansiForegrounds'
+export {default as ansiAll} from './ansi-all'
+export {default as ansiBackgrounds} from './ansi-backgrounds'
+export {default as ansiForegrounds} from './ansi-foregrounds'
 export {default as hex} from './hex'
 export {default as quasar} from './quasar'
 export {default as rgb} from './rgb'
